Removed tba v2 stuff and renamed tbav3 file
