+++ .gitignore
@@ -6,5 +6,3 @@
 /build
 /captures
 Scouting/Scouting.iml
-Scouting/Scouting.iml
-Scouting/Scouting.iml
